+++ core/src/main/java/com/jjg/game/core/service/ShopService.java
@@ -143,7 +143,7 @@ public class ShopService {
      * @param productId
      * @return
      */
-    public ShopProduct getShopProduct(int productId) {
+    public ShopProduct getShopProduct(long productId) {
         if (this.shopProductMap == null || this.shopProductMap.isEmpty()) {
             return null;
         }
